apps/backend/requirements.txt
Django==5.2.5
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
coverage==7.3.2

# Production Dependencies
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
django-prometheus==2.3.1
